- iceeey used std::string in place of char* for storing strings and     std::vector in place of Array for storing arrays.     Using std::string allows us to remove a lot of string copying code.   - iceeey fixed the numerous places in usermessages.h and the other     message classes where a char* is cast to a const char*. Doing this     cast can lead to a memory leak.   - iceeey removed Array type.   - iceeey added escape function to Database for escaping strings before     using them in a database query.   - iceeey converted database tables to use std::ostringstream for     creating queries.   - iceeey made a whole lot of get functions const.

+++ src/common/network/udp/udpsocket.cpp
@@ -84,7 +84,7 @@ Socket::~Socket()
 #endif
 }
 
-size_t Socket::receive(char* msg, size_t length, SocketAddress* sockaddr)
+int Socket::receive(char* msg, size_t length, SocketAddress* sockaddr)
 {
   sockaddr_in addr;
   socklen len = sizeof(addr);
